Clean up code for all packages but 'constraints'

+++ solver/src/main/java/org/chocosolver/solver/search/loop/lns/neighbors/PropagationGuidedNeighborhood.java
@@ -13,16 +13,11 @@ import org.chocosolver.solver.Model;
 import org.chocosolver.solver.Solution;
 import org.chocosolver.solver.exception.ContradictionException;
 import org.chocosolver.solver.variables.IntVar;
+import org.chocosolver.util.tools.MathUtils;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.BitSet;
-import java.util.Comparator;
-import java.util.List;
-import java.util.Random;
+import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import org.chocosolver.util.tools.MathUtils;
 
 /**
  * A Propagation Guided LNS
@@ -145,7 +140,7 @@ public class PropagationGuidedNeighborhood extends IntNeighbor {
                 candidates = IntStream.range(0, n)
                         .filter(i -> fragment.get(i) && all[i] > 0)
                         .boxed()
-                        .sorted(Comparator.comparingInt(i -> -all[(int)i]))
+                        .sorted(Comparator.comparingInt(i -> -all[i]))
                         .limit(listSize)
                         .collect(Collectors.toList());
             } else {
