+++ solver/src/test/java/org/chocosolver/solver/constraints/nary/RegularTest.java
@@ -224,10 +224,10 @@ public class RegularTest {
         }
 
         assertEquals(1, solutions.size());
-        assertEquals(-9, (int) solutions.get(0).getIntVal(CS[0]));
-        assertEquals(1, (int) solutions.get(0).getIntVal(CS[1]));
-        assertEquals(1, (int) solutions.get(0).getIntVal(CS[2]));
-        assertEquals(-5, (int) solutions.get(0).getIntVal(CS[3]));
+        assertEquals(-9, solutions.get(0).getIntVal(CS[0]));
+        assertEquals(1, solutions.get(0).getIntVal(CS[1]));
+        assertEquals(1, solutions.get(0).getIntVal(CS[2]));
+        assertEquals(-5, solutions.get(0).getIntVal(CS[3]));
     }
 
     @Test(groups="1s", timeOut=60000)
